Length bin functionality 202004

See merge request fisheries/modelling/casal2-development!34

+++ CASAL2/source/DerivedQuantities/Length/Biomass.cpp
@@ -24,7 +24,7 @@ namespace length {
  */
 void Biomass::PreExecute() {
   cache_value_ = 0.0;
-  vector<unsigned> length_bins = model_->length_bins();
+  vector<double> length_bins = model_->length_bins();
   auto iterator = partition_.begin();
   unsigned time_step_index = model_->managers().time_step()->current_time_step();
   LOG_FINE() << "Time step for calculating biomass = " << time_step_index;
@@ -54,7 +54,7 @@ void Biomass::PreExecute() {
 void Biomass::Execute() {
   LOG_TRACE();
   Double value = 0.0;
-  vector<unsigned> length_bins = model_->length_bins();
+  vector<double> length_bins = model_->length_bins();
   unsigned time_step_index = model_->managers().time_step()->current_time_step();
   LOG_FINE() << "Time step for calculating biomass = " << time_step_index;
   if (model_->state() == State::kInitialise) {
