+++ CASAL2/source/Selectivities/Common/AllValues.Test.cpp
@@ -70,7 +70,7 @@ TEST(Selectivities, AllValues_Age) {
  */
 TEST(Selectivities, AllValues_Length) {
   MockModel model;
-  vector<unsigned> lengths = {10, 20, 30, 40, 50, 60, 120};
+  vector<double> lengths = {10, 20, 30, 40, 50, 60, 120};
   EXPECT_CALL(model, min_age()).WillRepeatedly(Return(10));
   EXPECT_CALL(model, max_age()).WillRepeatedly(Return(20));
   EXPECT_CALL(model, age_spread()).WillRepeatedly(Return(11));
