+++ CASAL2/source/Selectivities/Common/Increasing.cpp
@@ -69,10 +69,11 @@ void Increasing::DoValidate() {
         << "Expected: " << (high_ - low_ + 1) << ", parsed " << v_.size();
     }
   }   else if (model_->partition_type() == PartitionType::kLength) {
-    vector<unsigned> length_bins = model_->length_bins();
+    vector<double> length_bins = model_->length_bins();
     if (low_ < length_bins[0] || low_ > length_bins[length_bins.size()-1])
       LOG_ERROR_P(PARAM_L) << ": 'l' (" << low_ << ") must be between the model min length (" << length_bins[0]
         << ") and max length (" << length_bins[length_bins.size()-1] << ")";
+
     unsigned bins = 0;
     for (unsigned i = 0; i < length_bins.size(); ++i) {
       if (length_bins[i] >= low_ && length_bins[i] <= high_)
@@ -115,15 +116,17 @@ void Increasing::RebuildCache() {
       }
     }
   } else if (model_->partition_type() == PartitionType::kLength) {
-    vector<unsigned> length_bins = model_->length_bins();
+    vector<double> length_bins = model_->length_bins();
     unsigned mark = 0;
     unsigned start_element = 0;
+
     while (mark == 0) {
       ++start_element;
       if (length_bins[start_element] >= low_) {
         ++mark;
       }
     }
+
     for (unsigned length_bin_index = 0; length_bin_index < length_bins.size(); ++length_bin_index) {
       Double temp = (Double)length_bins[length_bin_index];
       if (temp < low_) {
