+++ CASAL2/source/Selectivities/Common/InverseLogistic.Test.cpp
@@ -65,7 +65,7 @@ TEST(Selectivities, InverseLogistic_Length) {
   EXPECT_CALL(model, min_age()).WillRepeatedly(Return(10));
   EXPECT_CALL(model, max_age()).WillRepeatedly(Return(20));
   EXPECT_CALL(model, age_spread()).WillRepeatedly(Return(11));
-  vector<unsigned> lengths = {10, 20, 30, 40, 50, 60, 120};
+  vector<double> lengths = {10, 20, 30, 40, 50, 60, 120};
   EXPECT_CALL(model, length_bins()).WillRepeatedly(ReturnRef(lengths));
   EXPECT_CALL(model, partition_type()).WillRepeatedly(Return(PartitionType::kLength));
