+++ CASAL2/source/Selectivities/Common/KnifeEdge.cpp
@@ -53,7 +53,7 @@ void KnifeEdge::RebuildCache() {
         values_[age - age_index_] = 0.0;
     }
   } else if (model_->partition_type() == PartitionType::kLength) {
-    vector<unsigned> length_bins = model_->length_bins();
+    vector<double> length_bins = model_->length_bins();
 
     for (unsigned length_bin_index = 0; length_bin_index < length_bins.size(); ++length_bin_index) {
       Double temp = (Double)length_bins[length_bin_index];
