+++ CASAL2/source/Selectivities/Common/Logistic.cpp
@@ -77,7 +77,7 @@ void Logistic::RebuildCache() {
     }
   } else if (model_->partition_type() == PartitionType::kLength) {
     Double threshold = 0.0;
-    vector<unsigned> length_bins = model_->length_bins();
+    vector<double> length_bins = model_->length_bins();
 
     for (unsigned length_bin_index = 0; length_bin_index < length_bins.size(); ++length_bin_index) {
       Double temp = (Double)length_bins[length_bin_index];
